tools/requirements.txt
packaging==24.1
psycopg2-binary==2.9.9
StrEnum==0.4.15
